Break out modules for isolated concerns and add separate CI jobs (#302)
* Make examples its own module

* Make cosmos specific module and run in CI in separate task

* split up modules for separate CI jobs

* Use self hosted for unit tests also

* go mod tidy

* Add go mod tidy diff from runner

* move builtin chains to yaml file (#294)

* move builtin chains to yaml file

* fall back to embeded configuredChains.yaml

* spelling

* please the linter

* remove comma

* hyphenate bech32prefix

* check environment variable and nits

* comment

* logger

* remove logger from ChainSpec

* handle feedback

* bump setup-go and checkout to v3
