+++ .github/workflows/tests.yml
@@ -11,25 +11,18 @@ on:
       - '**.sum'
 
 jobs:
-  test:
-    name: tests
+  test-unit:
+    name: unit-tests
     runs-on: [self-hosted, linux]
     steps:
       # Install and setup go
       - name: Set up Go 1.18
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v3
         with:
           go-version: 1.18
 
       - name: checkout ibctest
-        uses: actions/checkout@v2
-
-      # cleanup docker environment on self-hosted test runner
-      # - name: prepare fresh docker environment
-      #   run: |
-      #     docker stop $(docker ps -a -q) || true
-      #     docker rm -f $(docker ps -a -q) || true
-      #     docker network prune -f || true
+        uses: actions/checkout@v3
 
       # cleanup environment on self-hosted test runner
       - name: clean
@@ -37,6 +30,68 @@ jobs:
           rm -rf ~/.ibctest
 
       # run tests
-      - name: run all tests
-        # show cosmos chain and relayer logs on test failure
-        run: (go test -race -timeout 30m -v -p 2 ./...) || (echo "\n\n*****CHAIN and RELAYER LOGS*****" && cat "$HOME/.ibctest/logs/ibctest.log" && exit 1)
+      - name: run unit tests
+        run: go test -race -timeout 30m -v -p 2 ./...
+  test-cmd-ibctest:
+    name: test-cmd-ibctest
+    runs-on: [self-hosted, linux]
+    steps:
+      # Install and setup go
+      - name: Set up Go 1.18
+        uses: actions/setup-go@v3
+        with:
+          go-version: 1.18
+
+      - name: checkout ibctest
+        uses: actions/checkout@v3
+
+      # cleanup environment on self-hosted test runner
+      - name: clean
+        run: |-
+          rm -rf ~/.ibctest
+
+      # run tests
+      - name: run conformance tests
+        run: (cd cmd/ibctest && go test -race -timeout 30m -v -p 2 ./...) || (echo "\n\n*****CHAIN and RELAYER LOGS*****" && cat "$HOME/.ibctest/logs/ibctest.log" && exit 1)
+  test-ibc-examples:
+    name: test-ibc-examples
+    runs-on: [self-hosted, linux]
+    steps:
+      # Install and setup go
+      - name: Set up Go 1.18
+        uses: actions/setup-go@v3
+        with:
+          go-version: 1.18
+
+      - name: checkout ibctest
+        uses: actions/checkout@v3
+
+      # cleanup environment on self-hosted test runner
+      - name: clean
+        run: |-
+          rm -rf ~/.ibctest
+
+      # run tests
+      - name: run example ibc tests
+        run: cd examples/ibc && go test -race -timeout 30m -v -p 2 ./...
+  test-cosmos-examples:
+    name: test-cosmos-examples
+    runs-on: [self-hosted, linux]
+    steps:
+      # Install and setup go
+      - name: Set up Go 1.18
+        uses: actions/setup-go@v3
+        with:
+          go-version: 1.18
+
+      - name: checkout ibctest
+        uses: actions/checkout@v3
+
+      # cleanup environment on self-hosted test runner
+      - name: clean
+        run: |-
+          rm -rf ~/.ibctest
+
+      # run tests
+      - name: run example cosmos tests
+        run: cd examples/cosmos && go test -race -timeout 30m -v -p 2 ./...
